Add files via upload
I have built a front end of an e-commerce store. It's basically built using HTML, Css, and JavaScript. Frontend is what appears on the user’s side most and what entices the users to give the website a stroll. Hence, the designers work at the base level of this process. So to make the design more attractive, I have used bootstrap also in it.